Reorganized Folder, Templates and Kanban Board
Always wanted an offline kanban board, so I setup the kanban board inside of obsidian. Created a sample task and a template for future tasks that you will create when populating the kanban board. Aside from this I reorganized all the folders in my Vault to group common topics into a single folder.